Added to resources folder.
As config files, files, etc, are put in the "resources" file, so it's less confusing for other users. The main files go in the github page, such as: src folder, plugin.yml, README.MD, and .poggit.yml file (If using poggit). And the files which go in the resources file are the config files.